+++ packages/content-blocks-kit/src/Form/Type/TableRowType.php
@@ -16,6 +16,7 @@ final class TableRowType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder->add('cells', LiveCollectionType::class, [
+            'cb_open_entries' => 'none',
             'label' => 'cb_kit.block.table.field.cells',
             'translation_domain' => 'content_blocks_kit',
             'entry_type' => TableCellType::class,